evaaa_train/requirements.txt
mlagents_envs==1.1.0
numpy==1.23.5
torch==2.5.0
hydra-core==1.3.0
lightning==2.5.0
tensorboard==2.18.0
rich==13.9.4
gymnasium==1.0.0
opencv-python==4.10.0.84